config: move `permissions.py` to `metime` package
